+++ src/components/mod.rs
@@ -2,7 +2,9 @@ pub mod boxed;
 pub mod divider;
 pub mod label;
 pub mod list;
+pub mod note;
 pub mod section;
+pub mod table;
 pub mod tag;
 
 pub use boxed::Boxed;
@@ -10,4 +12,6 @@ pub use divider::Divider;
 pub use label::Label;
 pub use list::List;
 pub use section::Section;
+pub use table::Table;
 pub use tag::Tag;
+pub use note::Note;