+++ src/lib.rs
@@ -12,9 +12,13 @@ pub use components::Boxed;
 pub use components::Divider;
 /// Re-exports the `Label` struct from the `components` module.
 pub use components::Label;
-/// Re-exports the `Section` struct from the `components` module.
-pub use components::Section;
-/// Re-exports the `Tag` struct from the `components` module.
-pub use components::Tag;
 /// Re-exports the `List` struct from the `components` module.
 pub use components::List;
+/// Re-exports the `Note` struct from the `components` module.
+pub use components::Note;
+/// Re-exports the `Section` struct from the `components` module.
+pub use components::Section;
+/// Re-exports the `Table` struct from the `components` module.
+pub use components::Table;
+/// Re-exports the `Tag` struct from the `components` module.
+pub use components::Tag;